Add Progressive Tab Representation
Progressive Scrolling is that the line below the tab titles to
represent touch movements accurately and continuously. So it doesn’t
jump suddenly as previously.
I added a new public property to enable or disable the functionality
(enabled by default).
It was done by conforming to the UIScrollViewDelegate to get the
ContentOffset and then use it to move the line. The width also changes
dynamically to accommodate the next titles.

Also a minor bug was fixed that caused the line to be stuck in its
former position when changing orientation.